+++ src/classes/GraphQL.js
@@ -3,14 +3,22 @@ import path from 'path';
 import Sequelize from 'sequelize';
 
 import { config } from 'posti';
-
-import {
-  logStep,
-} from 'posti/dist/utils';
-
+import { logStep } from 'posti/dist/utils';
 import Database from 'posti/dist/classes/Sequelize';
 
 global.config = config;
+
+/**
+ * Enable logging.
+ * @param {Object} proc - Process.
+ * @returns {Boolean} - Whether logging is enabled or not.
+ */
+const enableLogging = (proc = process) => (
+  proc.env.log
+    // eslint-disable-next-line no-console
+    ? console[(!['true', true].includes(proc.env.log) ? proc.env.log : 'log')]
+    : false
+);
 
 /**
  * GraphQL
@@ -30,7 +38,7 @@ class GraphQL extends Database {
         global.config.user,
         global.config.password,
         {
-          logging: false,
+          logging: enableLogging(),
           host: global.config.host,
           dialect: global.config.dialect,
           ...global.config.dialectOptions,
@@ -82,20 +90,29 @@ class GraphQL extends Database {
 
     Object.keys(tableConfig.fields).forEach((field) => {
       if (tableConfig.fields[field].extraComment) {
-        contents += `\n  ### ${field} comments:\n`;
-        contents += `  # - ${tableConfig.fields[field].extraComment.replace(/\n/g, '\n  # - ')}\n`;
-        contents += '  ###';
+        contents += `\n  """\n ## ${field} comments:\n`;
+        contents += `    - ${tableConfig.fields[field].extraComment.replace(/\n/g, '\n    - ')}\n`;
+        contents += '  """';
       }
       contents += `\n  ${field}: ${this.castGraphQLType(tableConfig.fields[field].type)}`;
     });
 
+    let extendContents = '';
+    const extensionFile = path.resolve(`${__dirname}/../graphql/Extend/${tableConfig.graphqlQuery}.js`);
+    if (fs.existsSync(extensionFile)) {
+      // eslint-disable-next-line global-require
+      extendContents = require(extensionFile).default;
+    }
+
     let definitions = 'const typeDefinitions = `\n';
     definitions += `type ${tableConfig.graphqlQuery} {`;
     definitions += `  ${contents}\n`;
+    definitions += `${extendContents}`;
     definitions += '}\n\n';
 
     definitions += `input ${tableConfig.graphqlQuery}Input {`;
     definitions += `  ${contents}\n`;
+    definitions += `${extendContents}`;
     definitions += '}`;\n\n';
 
     const fileSuffix = 'export default typeDefinitions;\n';
@@ -116,6 +133,9 @@ class GraphQL extends Database {
       case 'integer': {
         return 'Int';
       }
+      case 'buildingNumber': {
+        return 'Int';
+      }
       case 'YYYYMMDD': {
         return 'Date';
       }
@@ -127,3 +147,6 @@ class GraphQL extends Database {
 }
 
 export default GraphQL;
+export {
+  enableLogging,
+};