+++ src/graphql/Scalars/index.js
@@ -1,5 +1,6 @@
 const typeDefinitions = `
 scalar Date
+scalar BuildingNumber
 `;
 
 export default typeDefinitions;