+++ src/graphql/Types/PostalCode.js
@@ -11,16 +11,17 @@ type PostalCode {
   postOfficeShortName: String
   postOfficeShortNameSwe: String
   entryIntoForceAt: Date
-  ### typeCode comments:
-  # - 1 = Normal postcode
-  # - 2 = PO Box postcode
-  # - 3 = Corporate postal code
-  # - 4 = Compilation postcode
-  # - 5 = Reply Mail postcode
-  # - 6 = SmartPOST (Parcel machine)
-  # - 7 = Pick-up Point postcode
-  # - 8 = Technical postcode
-  ###
+  """
+ ## typeCode comments:
+    - 1 = Normal postcode
+    - 2 = PO Box postcode
+    - 3 = Corporate postal code
+    - 4 = Compilation postcode
+    - 5 = Reply Mail postcode
+    - 6 = SmartPOST (Parcel machine)
+    - 7 = Pick-up Point postcode
+    - 8 = Technical postcode
+  """
   typeCode: Int
   regionId: String
   regionName: String
@@ -28,12 +29,13 @@ type PostalCode {
   municipalityIdCode: String
   municipalityName: String
   municipalityNameSwe: String
-  ### municipalityLanguage comments:
-  # - 1 = Finnish
-  # - 2 = Bilingual
-  # - 3 = Bilingual
-  # - 4 = Swedish
-  ###
+  """
+ ## municipalityLanguage comments:
+    - 1 = Finnish
+    - 2 = Bilingual
+    - 3 = Bilingual
+    - 4 = Swedish
+  """
   municipalityLanguage: Int
 }
 
@@ -47,16 +49,17 @@ input PostalCodeInput {
   postOfficeShortName: String
   postOfficeShortNameSwe: String
   entryIntoForceAt: Date
-  ### typeCode comments:
-  # - 1 = Normal postcode
-  # - 2 = PO Box postcode
-  # - 3 = Corporate postal code
-  # - 4 = Compilation postcode
-  # - 5 = Reply Mail postcode
-  # - 6 = SmartPOST (Parcel machine)
-  # - 7 = Pick-up Point postcode
-  # - 8 = Technical postcode
-  ###
+  """
+ ## typeCode comments:
+    - 1 = Normal postcode
+    - 2 = PO Box postcode
+    - 3 = Corporate postal code
+    - 4 = Compilation postcode
+    - 5 = Reply Mail postcode
+    - 6 = SmartPOST (Parcel machine)
+    - 7 = Pick-up Point postcode
+    - 8 = Technical postcode
+  """
   typeCode: Int
   regionId: String
   regionName: String
@@ -64,12 +67,13 @@ input PostalCodeInput {
   municipalityIdCode: String
   municipalityName: String
   municipalityNameSwe: String
-  ### municipalityLanguage comments:
-  # - 1 = Finnish
-  # - 2 = Bilingual
-  # - 3 = Bilingual
-  # - 4 = Swedish
-  ###
+  """
+ ## municipalityLanguage comments:
+    - 1 = Finnish
+    - 2 = Bilingual
+    - 3 = Bilingual
+    - 4 = Swedish
+  """
   municipalityLanguage: Int
 }`;