+++ src/graphql/Types/PostalCodeChange.js
@@ -26,14 +26,15 @@ type PostalCodeChange {
   regionName: String
   regionNameSwe: String
   changedAt: Date
-  ### eventCode comments:
-  # - 1 = Change of name
-  # - 2 = Postal code closed
-  # - 3 = New postal code
-  # - 4 = Postal code merged
-  # - 5 = Postal code reactivation
-  # - 6 = Postal code replaced by new postal code
-  ###
+  """
+ ## eventCode comments:
+    - 1 = Change of name
+    - 2 = Postal code closed
+    - 3 = New postal code
+    - 4 = Postal code merged
+    - 5 = Postal code reactivation
+    - 6 = Postal code replaced by new postal code
+  """
   eventCode: Int
 }
 
@@ -62,14 +63,15 @@ input PostalCodeChangeInput {
   regionName: String
   regionNameSwe: String
   changedAt: Date
-  ### eventCode comments:
-  # - 1 = Change of name
-  # - 2 = Postal code closed
-  # - 3 = New postal code
-  # - 4 = Postal code merged
-  # - 5 = Postal code reactivation
-  # - 6 = Postal code replaced by new postal code
-  ###
+  """
+ ## eventCode comments:
+    - 1 = Change of name
+    - 2 = Postal code closed
+    - 3 = New postal code
+    - 4 = Postal code merged
+    - 5 = Postal code reactivation
+    - 6 = Postal code replaced by new postal code
+  """
   eventCode: Int
 }`;